backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.4.0
pytz==2023.3
sqlparse==0.4.4
django-filter==23.3
reportlab==4.4.4
joblib>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
